feat: moved RAD adapter in from f-cell-options to f-cell
